+++ src/main/resources/META-INF/maven/archetype-metadata.xml
@@ -14,7 +14,7 @@ name="nexus-format-archetype">
       <defaultValue>nexus-repository-${pluginFormat}</defaultValue>
     </requiredProperty>
     <requiredProperty key="pluginClass">
-      <!-- @todo Default to capitalized version of ${pluginFormat} -->
+      <defaultValue>${capitalPluginFormat}</defaultValue>
     </requiredProperty>
     <requiredProperty key="version">
       <defaultValue>0.0.1-SNAPSHOT</defaultValue>
@@ -22,15 +22,38 @@ name="nexus-format-archetype">
   </requiredProperties>
 
   <fileSets>
-    <fileSet filtered="true" packaged="true">
-      <directory>src/main/java/</directory>
-    </fileSet>
-    <fileSet filtered="true" packaged="true">
-      <directory>src/test/java/</directory>
-    </fileSet>
-    <fileSet filtered="true">
-      <directory>src/main/resources</directory>
+    <fileSet filtered="true" packaged="false">
+      <directory/>
+      <includes>
+        <include>README.md</include>
+        <include>docs/*</include>
+      </includes>
     </fileSet>
   </fileSets>
 
+  <modules>
+    <module id="${rootArtifactId}" dir="__rootArtifactId__" name="${rootArtifactId}">
+      <fileSets>
+        <fileSet filtered="true" packaged="true">
+          <directory>src/main/java/</directory>
+        </fileSet>
+        <fileSet filtered="true" packaged="true">
+          <directory>src/test/java/</directory>
+        </fileSet>
+        <fileSet filtered="true">
+          <directory>src/main/resources</directory>
+        </fileSet>
+      </fileSets>
+    </module>
+    <module id="${rootArtifactId}-it" dir="__rootArtifactId__-it" name="${rootArtifactId}-it">
+      <fileSets>
+        <fileSet filtered="true" packaged="true">
+          <directory>src/test/java/</directory>
+        </fileSet>
+        <fileSet filtered="true">
+          <directory>src/test/resources</directory>
+        </fileSet>
+      </fileSets>
+    </module>
+  </modules>
 </archetype-descriptor>
