+++ src/main/resources/archetype-resources/__rootArtifactId__/src/main/java/__pluginFormat__/internal/__pluginClass__RecipeSupport.groovy
