+++ src/main/resources/archetype-resources/__rootArtifactId__/src/main/resources/static/rapture/NX/__pluginFormat__/app/PluginConfig.js
