+++ src/main/resources/archetype-resources/__rootArtifactId__/src/main/resources/static/rapture/NX/__pluginFormat__/util/__pluginClass__RepositoryUrls.js
