+++ src/main/resources/archetype-resources/__rootArtifactId__/src/main/resources/static/rapture/resources/__artifactId__-debug.css
