+++ src/main/resources/archetype-resources/pom.xml
@@ -1,4 +1,5 @@
 #set( $currentYear = $package.getClass().forName("java.util.Date").newInstance().getYear() + 1900 )
+#set( $capitalPluginFormat = $pluginFormat.substring(0, 1).toUpperCase() + $pluginFormat.substring(1) )
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
@@ -17,77 +18,86 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
+
   <parent>
     <groupId>org.sonatype.nexus.plugins</groupId>
     <artifactId>nexus-plugins</artifactId>
     <version>${nexusPluginsVersion}</version>
   </parent>
+
+  <artifactId>${artifactId}-parent</artifactId>
+
   <groupId>\${groupId}</groupId>
-  <artifactId>\${artifactId}</artifactId>
-  <name>\${project.groupId}:\${project.artifactId}</name>
-  <version>\${version}</version>
+  <version>${version}</version>
+  <packaging>pom</packaging>
+
   <inceptionYear>${currentYear}</inceptionYear>
-  <packaging>bundle</packaging>
 
-  <dependencies>
+  <modules>
+    <module>${rootArtifactId}</module>
+    <module>${rootArtifactId}-it</module>
+  </modules>
 
-    <dependency>
-      <groupId>org.sonatype.nexus</groupId>
-      <artifactId>nexus-plugin-api</artifactId>
-      <scope>provided</scope>
-    </dependency>
+  <distributionManagement>
+    <snapshotRepository>
+      <id>ossrh</id>
+      <url>https://oss.sonatype.org/content/repositories/snapshots</url>
+    </snapshotRepository>
+  </distributionManagement>
 
-    <dependency>
-      <groupId>org.sonatype.nexus</groupId>
-      <artifactId>nexus-repository</artifactId>
-      <scope>provided</scope>
-    </dependency>
+  <properties>
+    <nxrm-version>${nexusPluginsVersion}</nxrm-version>
+  </properties>
 
-    <dependency>
-      <groupId>org.apache.commons</groupId>
-      <artifactId>commons-compress</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>org.sonatype.goodies</groupId>
-      <artifactId>goodies-testsupport</artifactId>
-      <scope>test</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>org.sonatype.nexus</groupId>
-      <artifactId>nexus-rapture</artifactId>
-      <scope>provided</scope>
-    </dependency>
-
-  </dependencies>
+  <repositories>
+    <!-- ensure we can find the parent pom when starting from an empty local .m2 repository -->
+    <repository>
+      <id>rso-snapshots</id>
+      <url>https://repository.sonatype.org/content/repositories/snapshots</url>
+    </repository>
+  </repositories>
 
   <build>
     <plugins>
       <plugin>
-        <groupId>org.sonatype.nexus.buildsupport</groupId>
-        <artifactId>extjs-maven-plugin</artifactId>
+        <artifactId>maven-surefire-plugin</artifactId>
+        <version>2.20</version>
         <configuration>
-          <namespace>NX.${pluginFormat}</namespace>
+          <useSystemClassLoader>false</useSystemClassLoader>
         </configuration>
       </plugin>
 
       <plugin>
         <groupId>org.sonatype.plugins</groupId>
-        <artifactId>yuicompressor-maven-plugin</artifactId>
+        <artifactId>nexus-staging-maven-plugin</artifactId>
+        <version>1.6.7</version>
+        <extensions>true</extensions>
+        <configuration>
+          <serverId>rso</serverId>
+          <nexusUrl>https://repository.sonatype.org/</nexusUrl>
+          <autoReleaseAfterClose>true</autoReleaseAfterClose>
+        </configuration>
       </plugin>
 
       <plugin>
-        <groupId>org.apache.karaf.tooling</groupId>
-        <artifactId>karaf-maven-plugin</artifactId>
+        <artifactId>maven-gpg-plugin</artifactId>
+        <version>1.6</version>
+        <executions>
+          <execution>
+            <id>sign-artifacts</id>
+            <phase>deploy</phase>
+            <goals>
+              <goal>sign</goal>
+            </goals>
+            <configuration>
+              <gpgArguments>
+                <arg>--pinentry-mode</arg>
+                <arg>loopback</arg>
+              </gpgArguments>
+            </configuration>
+          </execution>
+        </executions>
       </plugin>
-
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-surefire-plugin</artifactId>
-        <version>2.20</version>
-      </plugin>
-
     </plugins>
   </build>
 </project>
