--- src/main/resources/archetype-resources/src/main/java/__pluginFormat__/internal/security/__pluginClass__FormatSecurityContributor.java
@@ -1,36 +1,0 @@
-/*
- * Sonatype Nexus (TM) Open Source Version
- * Copyright (c) ${currentYear}-present Sonatype, Inc.
- * All rights reserved. Includes the third-party code listed at http://links.sonatype.com/products/nexus/oss/attributions.
- *
- * This program and the accompanying materials are made available under the terms of the Eclipse Public License Version 1.0,
- * which accompanies this distribution and is available at http://www.eclipse.org/legal/epl-v10.html.
- *
- * Sonatype Nexus (TM) Professional Version is available from Sonatype, Inc. "Sonatype" and "Sonatype Nexus" are trademarks
- * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
- * Eclipse Foundation. All other trademarks are the property of their respective owners.
- */
-package ${package}.${pluginFormat}.internal.security;
-
-import javax.inject.Inject;
-import javax.inject.Named;
-import javax.inject.Singleton;
-
-import ${package}.${pluginFormat}.internal.${pluginClass}Format;
-
-import org.sonatype.nexus.repository.Format;
-import org.sonatype.nexus.repository.security.RepositoryFormatSecurityContributor;
-
-/**
- * ${pluginClass} format security resource.
- */
-@Named
-@Singleton
-public class ${pluginClass}FormatSecurityContributor
-    extends RepositoryFormatSecurityContributor
-{
-  @Inject
-  public ${pluginClass}FormatSecurityContributor(@Named(${pluginClass}Format.NAME) final Format format) {
-    super(format);
-  }
-}
