+++ src/test/resources/projects/it1/archetype.properties
@@ -1,6 +1,6 @@
 nexusPluginsVersion=${defaultNexusPluginsVersion}
-groupId=org.sonatype.nexus.repository
-package=org.sonatype.nexus.repository
+groupId=org.sonatype.nexus.plugins
+package=org.sonatype.nexus.plugins
 artifactId=nexus-repository-foo
 pluginFormat=foo
 pluginClass=Foo
