+++ src/test/resources/projects/it1/reference/nexus-repository-foo-it/src/test/java/org/sonatype/nexus/plugins/foo/internal/FooProxyIT.java
@@ -1,0 +1,80 @@
+/*
+ * Sonatype Nexus (TM) Open Source Version
+ * Copyright (c) 2019-present Sonatype, Inc.
+ * All rights reserved. Includes the third-party code listed at http://links.sonatype.com/products/nexus/oss/attributions.
+ *
+ * This program and the accompanying materials are made available under the terms of the Eclipse Public License Version 1.0,
+ * which accompanies this distribution and is available at http://www.eclipse.org/legal/epl-v10.html.
+ *
+ * Sonatype Nexus (TM) Professional Version is available from Sonatype, Inc. "Sonatype" and "Sonatype Nexus" are trademarks
+ * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
+ * Eclipse Foundation. All other trademarks are the property of their respective owners.
+ */
+package org.sonatype.nexus.plugins.foo.internal;
+
+import org.sonatype.goodies.httpfixture.server.fluent.Server;
+import org.sonatype.nexus.pax.exam.NexusPaxExamSupport;
+import org.sonatype.nexus.repository.Repository;
+import org.sonatype.nexus.repository.http.HttpStatus;
+import org.sonatype.nexus.testsuite.testsupport.FormatClientSupport;
+import org.sonatype.nexus.testsuite.testsupport.NexusITSupport;
+
+import org.hamcrest.MatcherAssert;
+import org.junit.Test;
+import org.ops4j.pax.exam.Configuration;
+import org.ops4j.pax.exam.Option;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
+import static org.sonatype.goodies.httpfixture.server.fluent.Behaviours.content;
+import static org.sonatype.goodies.httpfixture.server.fluent.Behaviours.error;
+import static org.sonatype.nexus.testsuite.testsupport.FormatClientSupport.status;
+
+public class FooProxyIT
+    extends FooITSupport
+{
+  private static final String TEST_PATH = "some/valid/path/for/foo/remote/repo";
+
+  private FooClient proxyClient;
+
+  private Repository proxyRepo;
+
+  @Configuration
+  public static Option[] configureNexus() {
+    return NexusPaxExamSupport.options(
+        NexusITSupport.configureNexusBase(),
+        nexusFeature("org.sonatype.nexus.plugins", "nexus-repository-foo")
+    );
+  }
+
+  @Test
+  public void unresponsiveRemoteProduces404() throws Exception {
+    Server server = Server.withPort(0).serve("/*")
+        .withBehaviours(error(HttpStatus.NOT_FOUND))
+        .start();
+    try {
+      proxyRepo = repos.createFooProxy("foo-test-proxy-notfound", server.getUrl().toExternalForm());
+      proxyClient = fooClient(proxyRepo);
+      MatcherAssert.assertThat(FormatClientSupport.status(proxyClient.get(TEST_PATH)), is(HttpStatus.NOT_FOUND));
+    }
+    finally {
+      server.stop();
+    }
+  }
+
+  @Test
+  public void retrieveFooWhenRemoteOffline() throws Exception {
+    Server server = Server.withPort(0).serve("/*")
+        .withBehaviours(content("Response"))
+        .start();
+    try {
+      proxyRepo = repos.createFooProxy("foo-test-proxy-offline", server.getUrl().toExternalForm());
+      proxyClient = fooClient(proxyRepo);
+      proxyClient.get(TEST_PATH);
+    }
+    finally {
+      server.stop();
+    }
+    assertThat(status(proxyClient.get(TEST_PATH)), is(200));
+  }
+}
