+++ src/test/resources/projects/it1/reference/nexus-repository-foo/src/main/java/org/sonatype/nexus/plugins/foo/internal/FooRecipeSupport.groovy
@@ -1,0 +1,108 @@
+/*
+ * Sonatype Nexus (TM) Open Source Version
+ * Copyright (c) ${ITReferenceYear}-present Sonatype, Inc.
+ * All rights reserved. Includes the third-party code listed at http://links.sonatype.com/products/nexus/oss/attributions.
+ *
+ * This program and the accompanying materials are made available under the terms of the Eclipse Public License Version 1.0,
+ * which accompanies this distribution and is available at http://www.eclipse.org/legal/epl-v10.html.
+ *
+ * Sonatype Nexus (TM) Professional Version is available from Sonatype, Inc. "Sonatype" and "Sonatype Nexus" are trademarks
+ * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
+ * Eclipse Foundation. All other trademarks are the property of their respective owners.
+ */
+package org.sonatype.nexus.plugins.foo.internal
+
+import javax.inject.Inject
+import javax.inject.Provider
+
+import org.sonatype.nexus.plugins.foo.internal.security.FooSecurityFacet
+
+import org.sonatype.nexus.repository.Format
+import org.sonatype.nexus.repository.RecipeSupport
+import org.sonatype.nexus.repository.Type
+import org.sonatype.nexus.repository.attributes.AttributesFacet
+import org.sonatype.nexus.repository.cache.NegativeCacheFacet
+import org.sonatype.nexus.repository.cache.NegativeCacheHandler
+import org.sonatype.nexus.repository.http.PartialFetchHandler
+import org.sonatype.nexus.repository.httpclient.HttpClientFacet
+import org.sonatype.nexus.repository.purge.PurgeUnusedFacet
+import org.sonatype.nexus.repository.search.SearchFacet
+import org.sonatype.nexus.repository.security.SecurityHandler
+import org.sonatype.nexus.repository.storage.DefaultComponentMaintenanceImpl
+import org.sonatype.nexus.repository.storage.StorageFacet
+import org.sonatype.nexus.repository.storage.UnitOfWorkHandler
+import org.sonatype.nexus.repository.view.ConfigurableViewFacet
+import org.sonatype.nexus.repository.view.handlers.BrowseUnsupportedHandler
+import org.sonatype.nexus.repository.view.handlers.ConditionalRequestHandler
+import org.sonatype.nexus.repository.view.handlers.ContentHeadersHandler
+import org.sonatype.nexus.repository.view.handlers.ExceptionHandler
+import org.sonatype.nexus.repository.view.handlers.HandlerContributor
+import org.sonatype.nexus.repository.view.handlers.TimingHandler
+
+/**
+ * Support for Foo recipes.
+ */
+abstract class FooRecipeSupport
+    extends RecipeSupport
+{
+  @Inject
+  Provider<FooSecurityFacet> securityFacet
+
+  @Inject
+  Provider<ConfigurableViewFacet> viewFacet
+
+  @Inject
+  Provider<StorageFacet> storageFacet
+
+  @Inject
+  Provider<SearchFacet> searchFacet
+
+  @Inject
+  Provider<AttributesFacet> attributesFacet
+
+  @Inject
+  ExceptionHandler exceptionHandler
+
+  @Inject
+  TimingHandler timingHandler
+
+  @Inject
+  SecurityHandler securityHandler
+
+  @Inject
+  PartialFetchHandler partialFetchHandler
+
+  @Inject
+  ConditionalRequestHandler conditionalRequestHandler
+
+  @Inject
+  ContentHeadersHandler contentHeadersHandler
+
+  @Inject
+  UnitOfWorkHandler unitOfWorkHandler
+
+  @Inject
+  BrowseUnsupportedHandler browseUnsupportedHandler
+
+  @Inject
+  HandlerContributor handlerContributor
+
+  @Inject
+  Provider<DefaultComponentMaintenanceImpl> componentMaintenanceFacet
+
+  @Inject
+  Provider<HttpClientFacet> httpClientFacet
+
+  @Inject
+  Provider<PurgeUnusedFacet> purgeUnusedFacet
+
+  @Inject
+  Provider<NegativeCacheFacet> negativeCacheFacet
+
+  @Inject
+  NegativeCacheHandler negativeCacheHandler
+
+  protected FooRecipeSupport(final Type type, final Format format) {
+    super(type, format)
+  }
+}
