+++ src/test/resources/projects/it1/reference/nexus-repository-foo/src/main/resources/static/rapture/NX/foo/app/PluginStrings.js
