+++ src/test/resources/projects/it1/reference/nexus-repository-foo/src/main/resources/static/rapture/resources/nexus-repository-foo-debug.css
