--- src/test/resources/projects/it1/reference/src/test/java/org/sonatype/nexus/repository/foo/internal/security/FooSecurityFacetTest.java
@@ -1,90 +1,0 @@
-/*
- * Sonatype Nexus (TM) Open Source Version
- * Copyright (c) ${ITReferenceYear}-present Sonatype, Inc.
- * All rights reserved. Includes the third-party code listed at http://links.sonatype.com/products/nexus/oss/attributions.
- *
- * This program and the accompanying materials are made available under the terms of the Eclipse Public License Version 1.0,
- * which accompanies this distribution and is available at http://www.eclipse.org/legal/epl-v10.html.
- *
- * Sonatype Nexus (TM) Professional Version is available from Sonatype, Inc. "Sonatype" and "Sonatype Nexus" are trademarks
- * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
- * Eclipse Foundation. All other trademarks are the property of their respective owners.
- */
-package org.sonatype.nexus.repository.foo.internal.security;
-
-import org.sonatype.goodies.testsupport.TestSupport;
-import org.sonatype.nexus.repository.Format;
-import org.sonatype.nexus.repository.Repository;
-import org.sonatype.nexus.repository.http.HttpMethods;
-import org.sonatype.nexus.repository.security.ContentPermissionChecker;
-import org.sonatype.nexus.repository.security.VariableResolverAdapter;
-import org.sonatype.nexus.repository.view.Request;
-
-import org.apache.shiro.authz.AuthorizationException;
-import org.junit.Before;
-import org.junit.Test;
-import org.mockito.Mock;
-
-import static org.junit.Assert.fail;
-import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.eq;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import static org.sonatype.nexus.security.BreadActions.READ;
-
-public class FooSecurityFacetTest
-    extends TestSupport
-{
-  @Mock
-  Request request;
-
-  @Mock
-  Repository repository;
-
-  @Mock
-  ContentPermissionChecker contentPermissionChecker;
-
-  @Mock
-  VariableResolverAdapter variableResolverAdapter;
-
-  @Mock
-  FooFormatSecurityContributor securityContributor;
-
-  FooSecurityFacet fooSecurityFacet;
-
-  @Before
-  public void setupConfig() throws Exception {
-    when(request.getPath()).thenReturn("/some/path.txt");
-    when(request.getAction()).thenReturn(HttpMethods.GET);
-
-    when(repository.getFormat()).thenReturn(new Format("foo") { });
-    when(repository.getName()).thenReturn("FooSecurityFacetTest");
-
-    fooSecurityFacet = new FooSecurityFacet(securityContributor,
-        variableResolverAdapter, contentPermissionChecker);
-
-    fooSecurityFacet.attach(repository);
-  }
-
-  @Test
-  public void testEnsurePermitted_permitted() throws Exception {
-    when(contentPermissionChecker.isPermitted(eq("FooSecurityFacetTest"), eq("foo"), eq(READ), any()))
-        .thenReturn(true);
-    fooSecurityFacet.ensurePermitted(request);
-  }
-
-  @Test
-  public void testEnsurePermitted_notPermitted() throws Exception {
-    when(contentPermissionChecker.isPermitted(eq("FooSecurityFacetTest"), eq("foo"), eq(READ), any()))
-        .thenReturn(false);
-    try {
-      fooSecurityFacet.ensurePermitted(request);
-      fail("AuthorizationException should have been thrown");
-    }
-    catch (AuthorizationException e) {
-      //expected
-    }
-
-    verify(contentPermissionChecker).isPermitted(eq("FooSecurityFacetTest"), eq("foo"), eq(READ), any());
-  }
-}
